Workaround a specific compiler rint() issues

+++ onnxruntime/core/providers/cpu/math/round.cc
@@ -36,7 +36,7 @@ Status Round<MLFloat16>::Compute(OpKernelContext* ctx) const {
   auto* output = Y.MutableData<MLFloat16>();
   const auto size = X.Shape().Size();
   for (int64_t i = 0; i < size; ++i, ++output, ++input) {
-    *output = MLFloat16(::rint(input->ToFloat()));
+    *output = MLFloat16(static_cast<float>(::rint(input->ToFloat())));
   }
   return Status::OK();
 }
